corrected error on migration name

+++ database/migrations/2022_03_03_123202_create_achievements_table.php
@@ -17,7 +17,7 @@ class CreateAchievementsTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('type');
-            $table->int('value');
+            $table->tinyInteger('value');
             $table->timestamps();
         });
