+++ database/migrations/2022_03_03_133925_create_badges_table.php
@@ -16,7 +16,7 @@ class CreateBadgesTable extends Migration
         Schema::create('badges', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->int('value');
+            $table->tinyInteger('value');
             $table->timestamps();
         });
